Pass encoding size to sequence reducer
If the agreement isn't passed along the `attention` reducer option doesn't work.
Note: I've only tested that this works for myself but looking at the code there's quite a few other places this would have to be applied. However I'm not confident I fully understand why it looks the way it currently does so thought that instead of just filing a bug I'd at least start a PR for it and hopefully someone else can chime in with what else I should consider.